+++ HIRS_AttestationCAPortal/config/checkstyle/suppressions.xml
@@ -6,4 +6,5 @@
 
 <suppressions>
     <suppress checks="MagicNumber" files=".*[/\\]src[/\\]test[/\\]+" />
+    <suppress checks="AvoidInlineConditionals" files=".*[/\\]src[/\\]test[/\\]+" />
 </suppressions>
